alter remainingTime type (int -> float)

+++ Assets/Script/GameManager.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class GameManager : MonoBehaviour {
 	public string handedness = "right";
 	public int score = 0;
-	public int remainingTime = 180; // sec
+	public float remainingTime = 180; // sec
 
 	// Use this for initialization
 	void Start() {
